Repositories refactor
